# docs/openapi/mescopilot.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MesCopilot.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/audit/operations": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDtoPagedAuditResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDtoPagedAuditResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDtoPagedAuditResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/changes": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataChangeLogDtoPagedAuditResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataChangeLogDtoPagedAuditResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataChangeLogDtoPagedAuditResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/agent": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAuditLogDtoPagedAuditResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAuditLogDtoPagedAuditResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAuditLogDtoPagedAuditResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/report": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentAuditLogDto": {
        "required": [
          "durationMilliseconds",
          "id",
          "mode",
          "status",
          "timestamp",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/AgentMode"
          },
          "status": {
            "type": "string"
          },
          "toolName": {
            "type": "string",
            "nullable": true
          },
          "queryDigest": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "nullable": true
          },
          "discrepancies": {
            "type": "string",
            "nullable": true
          },
          "durationMilliseconds": {
            "type": "integer",
            "format": "int64"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AgentAuditLogDtoPagedAuditResultDto": {
        "required": [
          "items",
          "skip",
          "take",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentAuditLogDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "skip": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AgentMode": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "AuditLogDto": {
        "required": [
          "durationMilliseconds",
          "id",
          "method",
          "queryKeys",
          "routeTemplate",
          "statusCode",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "method": {
            "type": "string"
          },
          "routeTemplate": {
            "type": "string"
          },
          "queryKeys": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "durationMilliseconds": {
            "type": "integer",
            "format": "int64"
          },
          "requestBody": {
            "type": "string",
            "nullable": true
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "ipHash": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuditLogDtoPagedAuditResultDto": {
        "required": [
          "items",
          "skip",
          "take",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "skip": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DataChangeLogDto": {
        "required": [
          "changeType",
          "entityId",
          "entityType",
          "id",
          "newValues",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "changeType": {
            "type": "string"
          },
          "oldValues": {
            "type": "string",
            "nullable": true
          },
          "newValues": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataChangeLogDtoPagedAuditResultDto": {
        "required": [
          "items",
          "skip",
          "take",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataChangeLogDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "skip": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TenantProblemDetails": {
        "required": [
          "code",
          "status",
          "title"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}